+++ tests/conftest.py
@@ -11,7 +11,12 @@ def redis_socket(tmp_path):
 @pytest.fixture
 def redis_server(tmp_path, redis_socket):
     with TestProcess(
-        'redis-server', '--port', '0', '--save', '', '--appendonly', 'yes', '--dir', tmp_path, '--unixsocket', redis_socket
+        'valkey-server', '--port', '0', '--save', '', '--appendonly', 'yes', '--dir', tmp_path, '--unixsocket', redis_socket
     ) as redis_server:
-        wait_for_strings(redis_server.read, 2, 'eady to accept connections')
+        wait_for_strings(
+            redis_server.read,
+            2,
+            'oO0OoO0OoO0Oo Valkey is starting oO0OoO0OoO0Oo',
+            'Ready to accept connections',
+        )
         yield redis_server